rename SentencePadding to SentenceBiPadding
