+++ dl/src/main/scala/com/intel/analytics/bigdl/dataset/text/SentenceBiPadding.scala
@@ -1,11 +1,12 @@
 /*
- * Copyright 2016 The BigDL Authors.
+ * Licensed to Intel Corporation under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * Intel Corporation licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *    http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -17,8 +18,6 @@
 package com.intel.analytics.bigdl.dataset.text
 
 import com.intel.analytics.bigdl.dataset.Transformer
-import com.intel.analytics.bigdl.dataset.text.utils.SentenceToken
-
 import scala.collection.Iterator
 
 /**
@@ -30,12 +29,12 @@ class SentenceBiPadding(
   end: Option[String] = None)
   extends Transformer[String, String] {
 
-  val sentenceStart = start.getOrElse(SentenceToken.start)
-  val sentenceEnd = end.getOrElse(SentenceToken.end)
+  val sentenceStart = start.getOrElse("SENTENCESTART ")
+  val sentenceEnd = end.getOrElse(" SENTENCEEND")
 
   override def apply(prev: Iterator[String]): Iterator[String] = {
     prev.map(x => {
-      val sentence = sentenceStart + " " + x + " " + sentenceEnd
+      val sentence = sentenceStart + x + sentenceEnd
       sentence
     })
   }
