--- dl/src/main/scala/com/intel/analytics/bigdl/dataset/text/SentencePadding.scala
@@ -1,47 +1,0 @@
-/*
- * Licensed to Intel Corporation under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * Intel Corporation licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.intel.analytics.bigdl.dataset.text
-
-import com.intel.analytics.bigdl.dataset.Transformer
-import scala.collection.Iterator
-
-/**
- * x =>  ["start", x, "end"]
- */
-
-class SentencePadding(
-  start: Option[String] = None,
-  end: Option[String] = None)
-  extends Transformer[String, String] {
-
-  val sentenceStart = start.getOrElse("SENTENCESTART ")
-  val sentenceEnd = end.getOrElse(" SENTENCEEND")
-
-  override def apply(prev: Iterator[String]): Iterator[String] = {
-    prev.map(x => {
-      val sentence = sentenceStart + x + sentenceEnd
-      sentence
-    })
-  }
-}
-
-object SentencePadding {
-  def apply(start: Option[String] = None,
-            end: Option[String] = None):
-  SentencePadding = new SentencePadding(start, end)
-}
